requirements.txt
# Core Dependencies
python-dotenv==1.1.0
langgraph==0.4.3
ollama==0.1.9
pydantic>=2.0.0,<3.0.0
pydantic-core>=2.16.2

# Vector Database and Embeddings for Memory
chromadb==0.4.23
sentence-transformers==2.5.0

# DSPy for Agent Cognitive Systems
dspy==2.6.23

# Implicit Dependencies (Discovered during analysis)
requests==2.32.3
numpy==2.2.5
colorama==0.4.6
tqdm==4.66.2
typing-extensions==4.9.0
packaging==23.2

# Discord Bot
discord.py==2.3.2

# Future Use (Currently unused but may be needed later)
# redis==5.0.1
# langchain==0.1.5
# langchain-community==0.0.13

# Add other dependencies as needed 
weaviate-client>=4.5.4
fastapi
uvicorn[standard]
sse-starlette 